Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@554 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ play/src/Player.java
@@ -72,7 +72,7 @@ public class Player implements Immutable
 		byte[] rb = new byte[0x100];
 		for (int i = 0; i < rb.length; i++)
         {
-            rb[i] = i;
+            rb[i] = (byte)i;
         }
 		ByteArrayInputStream bais = new ByteArrayInputStream(rb);
 		InputStreamReader isr = new InputStreamReader(bais,"US-ASCII");
